Grading
So I noticed this isn't the version that is up on shinyapps. I docked some points for that, also because you've only got code for two pages (one on shinyapps.io). I also removed a point because you've only got 2 pages, not three.

This took me an embarrasing amount of time to figure out, but its because you had the tabName spelled without capitals.